Main: VC++ 19.2 update 11 build adjustments (lossy floating point conversions). Also, IPRT does support formatting floating point numbers. bugref:8489
svn:sync-xref-src-repo-rev: r149378

+++ src/VBox/Main/testcase/tstCollector.cpp
@@ -1,4 +1,4 @@
-/* $Id: tstCollector.cpp 93115 2022-01-01 11:31:46Z knut.osmundsen@oracle.com $ */
+/* $Id: tstCollector.cpp 93310 2022-01-18 11:32:28Z knut.osmundsen@oracle.com $ */
 /** @file
  * VirtualBox Main - Performance collector classes test cases.
  */
@@ -142,7 +142,7 @@ void measurePerformance(pm::CollectorHAL *collector, const char *pszName, int cV
         /* Process RAM usage */
         N_CALLS(cVMs, getProcessMemoryUsage(processes[call], &tmp));
     }
-    RTPrintf("\n%d VMs -- %.2f%% of CPU time\n", cVMs, (RTTimeNanoTS() - start) / 10000000. / times);
+    RTPrintf("\n%d VMs -- %u%% of CPU time\n", cVMs, (unsigned)((double)(RTTimeNanoTS() - start) / 10000000.0 / times));
 
     /* Shut down fake VMs */
     shutdownProcessList(processes);
